requirements.txt
# Core ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.0.0

# Astronomy specific
lightkurve>=2.4.0
astropy>=5.3.0
astroquery>=0.4.6
transitleastsquares>=1.0.32

# Preprocessing
scipy>=1.11.0
numba>=0.57.0

# Imbalanced data
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Experiment tracking
tensorboard>=2.13.0
mlflow>=2.5.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.7.0
pytest>=7.4.0
flake8>=6.0.0